requirements.txt
certifi
click==8.1.3
colorama==0.4.6
Flask==2.2.3
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-decouple==3.7
Werkzeug==2.2.3
wincertstore==0.2
zipp==3.15.0
